+++ sdk/src/lib.rs
@@ -81,83 +81,26 @@ pub enum InputTapeType {
     PrivateTape,
 }
 
-/// Reads given number of raw bytes from an input tape
-#[allow(clippy::missing_errors_doc)]
-#[cfg(all(feature = "std", feature = "stdread", target_os = "mozakvm"))]
-pub fn read(kind: &InputTapeType, buf: &mut [u8]) -> std::io::Result<usize> {
-    use std::io::Read;
-    match kind {
-        InputTapeType::PublicTape => unsafe {
-            crate::common::system::SYSTEM_TAPE
-                .public_input_tape
-                .read(buf)
-        },
-        InputTapeType::PrivateTape => unsafe {
-            crate::common::system::SYSTEM_TAPE
-                .private_input_tape
-                .read(buf)
-        },
-    }
-}
-
-/// Reads given number of raw bytes from an input tape
-#[allow(clippy::missing_errors_doc)]
+/// Provides the length of tape available to read
 #[cfg(all(feature = "std", target_os = "mozakvm"))]
-#[must_use]
-pub fn input_tape_len(kind: &InputTapeType) -> usize {
-    match kind {
-        InputTapeType::PublicTape => unsafe {
-            crate::common::system::SYSTEM_TAPE.public_input_tape.len()
-        },
-        InputTapeType::PrivateTape => unsafe {
-            crate::common::system::SYSTEM_TAPE.private_input_tape.len()
-        },
-    }
-}
-
-/// Writes given number of raw bytes from an input tape
-#[allow(clippy::missing_errors_doc)]
-#[cfg(all(feature = "std", not(target_os = "mozakvm")))]
-pub fn write(kind: &InputTapeType, buf: &[u8]) -> std::io::Result<usize> {
-    use std::io::Write;
-    match kind {
-        InputTapeType::PublicTape => unsafe {
-            crate::common::system::SYSTEM_TAPE
-                .public_input_tape
-                .write(buf)
-        },
-        InputTapeType::PrivateTape => unsafe {
-            crate::common::system::SYSTEM_TAPE
-                .private_input_tape
-                .write(buf)
-        },
-    }
-}
-
-/// Manually add a `ProgramIdentifier` onto `IdentityStack`.
+pub use crate::mozakvm::inputtape::input_tape_len;
+/// Reads utmost given number of raw bytes from an input tape
+#[cfg(all(feature = "std", feature = "stdread", target_os = "mozakvm"))]
+pub use crate::mozakvm::inputtape::read;
+/// Manually add a `ProgramIdentifier` onto `IdentityStack`. Useful
+/// when one want to escape automatic management of `IdentityStack`
+/// via cross-program-calls sends (ideally temporarily).
 /// CAUTION: Manual function for `IdentityStack`, misuse may lead
 /// to system tape generation failure.
 #[cfg(all(feature = "std", not(target_os = "mozakvm")))]
-pub fn add_identity(id: crate::common::types::ProgramIdentifier) {
-    unsafe {
-        crate::common::system::SYSTEM_TAPE
-            .call_tape
-            .identity_stack
-            .borrow_mut()
-            .add_identity(id);
-    }
-}
-
+pub use crate::native::helpers::add_identity;
 /// Manually remove a `ProgramIdentifier` from `IdentityStack`.
+/// Useful when one want to escape automatic management of `IdentityStack`
+/// via cross-program-calls sends (ideally temporarily).
 /// CAUTION: Manual function for `IdentityStack`, misuse may lead
 /// to system tape generation failure.
 #[cfg(all(feature = "std", not(target_os = "mozakvm")))]
-pub fn rm_identity() {
-    unsafe {
-        crate::common::system::SYSTEM_TAPE
-            .call_tape
-            .identity_stack
-            .borrow_mut()
-            .rm_identity();
-    }
-}
+pub use crate::native::helpers::rm_identity;
+/// Writes raw bytes to an input tape. Infallible
+#[cfg(all(feature = "std", not(target_os = "mozakvm")))]
+pub use crate::native::inputtape::write;
