+++ sdk/src/mozakvm/inputtape.rs
@@ -189,3 +189,36 @@ impl PublicInputTape {
 
     pub fn read_ptr(&self) -> usize { self.0.read_ptr() }
 }
+
+/// Provides the length of tape available to read
+#[cfg(all(feature = "std", target_os = "mozakvm"))]
+#[must_use]
+pub fn input_tape_len(kind: &crate::InputTapeType) -> usize {
+    match kind {
+        crate::InputTapeType::PublicTape => unsafe {
+            crate::common::system::SYSTEM_TAPE.public_input_tape.len()
+        },
+        crate::InputTapeType::PrivateTape => unsafe {
+            crate::common::system::SYSTEM_TAPE.private_input_tape.len()
+        },
+    }
+}
+
+/// Reads utmost given number of raw bytes from an input tape
+#[allow(clippy::missing_errors_doc)]
+#[cfg(all(feature = "std", feature = "stdread", target_os = "mozakvm"))]
+pub fn read(kind: &crate::InputTapeType, buf: &mut [u8]) -> std::io::Result<usize> {
+    use std::io::Read;
+    match kind {
+        crate::InputTapeType::PublicTape => unsafe {
+            crate::common::system::SYSTEM_TAPE
+                .public_input_tape
+                .read(buf)
+        },
+        crate::InputTapeType::PrivateTape => unsafe {
+            crate::common::system::SYSTEM_TAPE
+                .private_input_tape
+                .read(buf)
+        },
+    }
+}
